// src/interfaces/movie.ts
export interface IMovie {
    id: number;
    category: number;
    status: string;
    name: string;
    num_files: number;
    size: number;
    seeders: number;
    leechers: number;
    username: string;
    added: number;
    descr: string;
    imdb: string;
    image: string;
    language: string;
    textlanguage: string;
    info_hash: string;
    magnet: string;
}